added noAWT and noAWTDuringStartup options The planner now runs with -XstartOnFirstThread on mac Updated to SimpleLibraryPlus 1.0.1 Added support for negative cooling values (WARNING: likely unstable) Changed overhaul SFR outputRatio from int to float Added more cluster outlines to overhaul SFR and MSR Fixed IC2 fuels in S'plodo-bot Incremented minor version

+++ src/multiblock/configuration/overhaul/fissionsfr/CoolantRecipe.java
@@ -6,8 +6,8 @@ public class CoolantRecipe{
     public String input;
     public String output;
     public int heat;
-    public int outputRatio;
-    public CoolantRecipe(String name, String input, String output, int heat, int outputRatio){
+    public float outputRatio;
+    public CoolantRecipe(String name, String input, String output, int heat, float outputRatio){
         this.name = name;
         this.input = input;
         this.output = output;
